+++ compiler/src/main/java/com/sankuai/waimai/router/compiler/BaseProcessor.java
@@ -1,9 +1,7 @@
 package com.sankuai.waimai.router.compiler;
 
-import static com.google.common.base.Charsets.UTF_8;
-
 import com.sankuai.waimai.router.interfaces.Const;
-import com.sankuai.waimai.router.service.ServiceImpl;
+import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.CodeBlock;
 import com.squareup.javapoet.JavaFile;
 import com.squareup.javapoet.MethodSpec;
@@ -12,15 +10,10 @@ import com.squareup.javapoet.TypeSpec;
 import com.sun.tools.javac.code.Symbol;
 import com.sun.tools.javac.code.Type;
 
-import java.io.BufferedWriter;
 import java.io.IOException;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.math.BigInteger;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 
 import javax.annotation.processing.AbstractProcessor;
@@ -28,13 +21,10 @@ import javax.annotation.processing.Filer;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.Modifier;
-import javax.lang.model.element.PackageElement;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.TypeMirror;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
-import javax.tools.FileObject;
-import javax.tools.StandardLocation;
 
 /**
  * Created by jzj on 2018/3/23.
@@ -54,8 +44,36 @@ public abstract class BaseProcessor extends AbstractProcessor {
         elements = processingEnvironment.getElementUtils();
     }
 
+    /**
+     * 从字符串获取TypeElement对象
+     */
+    public TypeElement typeElement(String className) {
+        return elements.getTypeElement(className);
+    }
+
+    /**
+     * 从字符串获取TypeMirror对象
+     */
     public TypeMirror typeMirror(String className) {
-        return elements.getTypeElement(className).asType();
+        return typeElement(className).asType();
+    }
+
+    /**
+     * 从字符串获取ClassName对象
+     */
+    public ClassName className(String className) {
+        return ClassName.get(typeElement(className));
+    }
+
+    /**
+     * 从字符串获取TypeName对象，包含Class的泛型信息
+     */
+    public TypeName typeName(String className) {
+        return TypeName.get(typeMirror(className));
+    }
+
+    public static String getClassName(TypeMirror typeMirror) {
+        return typeMirror == null ? "" : typeMirror.toString();
     }
 
     public boolean isSubType(TypeMirror type, String className) {
@@ -92,10 +110,6 @@ public abstract class BaseProcessor extends AbstractProcessor {
         return isConcreteType(element) && isSubType(element, typeMirror);
     }
 
-    public boolean isType(TypeMirror interfaces, String name) {
-        return interfaces != null && interfaces.toString().equals(name);
-    }
-
     public boolean isActivity(Element element) {
         return isConcreteSubType(element, Const.ACTIVITY_CLASS);
     }
@@ -108,36 +122,19 @@ public abstract class BaseProcessor extends AbstractProcessor {
         return isConcreteSubType(element, Const.URI_INTERCEPTOR_CLASS);
     }
 
-    public static String getClassName(TypeMirror typeMirror) {
-        return typeMirror == null ? "" : typeMirror.toString();
-    }
-
-    public void writeHandlerInitClass(CodeBlock code, String hash,
-            String genClassName, String handlerClassName, String interfaceName) {
+    public static String hash(String str) {
         try {
-            genClassName += Const.SPLITTER + hash;
-            MethodSpec methodSpec = MethodSpec.methodBuilder(Const.INIT_METHOD)
-                    .addModifiers(Modifier.PUBLIC)
-                    .returns(TypeName.VOID)
-                    .addParameter(TypeName.get(typeMirror(handlerClassName)), "handler")
-                    .addCode(code)
-                    .build();
-            TypeSpec typeSpec = TypeSpec.classBuilder(genClassName)
-                    .addSuperinterface(TypeName.get(typeMirror(interfaceName)))
-                    .addModifiers(Modifier.PUBLIC)
-                    .addMethod(methodSpec)
-                    .build();
-            JavaFile.builder(Const.GEN_PKG, typeSpec)
-                    .build()
-                    .writeTo(filer);
-            String fullImplName = Const.GEN_PKG + Const.DOT + genClassName;
-            String config = new ServiceImpl(null, fullImplName, false).toConfig();
-            writeInterfaceServiceFile(interfaceName, Collections.singletonList(config));
-        } catch (IOException e) {
-            e.printStackTrace();
+            MessageDigest md = MessageDigest.getInstance("MD5");
+            md.update(str.getBytes());
+            return new BigInteger(1, md.digest()).toString(16);
+        } catch (NoSuchAlgorithmException e) {
+            return Integer.toHexString(str.hashCode());
         }
     }
 
+    /**
+     * 创建Handler。格式：<code>"com.demo.TestActivity"</code> 或 <code>new TestHandler()</code>
+     */
     public CodeBlock buildHandler(boolean isActivity, Symbol.ClassSymbol cls) {
         CodeBlock.Builder b = CodeBlock.builder();
         if (isActivity) {
@@ -148,6 +145,9 @@ public abstract class BaseProcessor extends AbstractProcessor {
         return b.build();
     }
 
+    /**
+     * 创建Interceptors。格式：<code>, new Interceptor1(), new Interceptor2()</code>
+     */
     public CodeBlock buildInterceptors(List<? extends TypeMirror> interceptors) {
         CodeBlock.Builder b = CodeBlock.builder();
         if (interceptors != null && interceptors.size() > 0) {
@@ -163,71 +163,117 @@ public abstract class BaseProcessor extends AbstractProcessor {
         return b.build();
     }
 
-    public void writeInterfaceServiceFile(String interfaceName, Collection<String> lines) {
-        writeServiceFile(Const.SERVICE_PATH + interfaceName, lines);
-    }
-
-    public void writeServiceFile(String fileName, Collection<String> lines) {
-        if (isEmpty(fileName) || isEmpty(lines)) {
-            return;
-        }
+    /**
+     * 生成类似下面格式的HandlerInitClass，同时生成ServiceInitClass
+     * <pre>
+     * package com.sankuai.waimai.router.generated;
+     * public class UriRouter_RouterUri_xxx implements IUriAnnotationInit {
+     *     public void init(UriAnnotationHandler handler) {
+     *         handler.register("", "", "/login", "com.xxx.LoginActivity", false);
+     *         // ...
+     *     }
+     * }
+     * </pre>
+     *
+     * @param code             方法中的代码
+     * @param genClassName     生成class的SimpleClassName，形如 UriRouter_RouterUri_xxx
+     * @param handlerClassName Handler类名，例如 com.sankuai.waimai.router.common.UriAnnotationHandler
+     * @param interfaceName    接口名，例如 com.sankuai.waimai.router.common.IUriAnnotationInit
+     */
+    public void buildHandlerInitClass(CodeBlock code, String genClassName, String handlerClassName, String interfaceName) {
+        MethodSpec methodSpec = MethodSpec.methodBuilder(Const.INIT_METHOD)
+                .addModifiers(Modifier.PUBLIC)
+                .returns(TypeName.VOID)
+                .addParameter(className(handlerClassName), "handler")
+                .addCode(code)
+                .build();
+        TypeSpec typeSpec = TypeSpec.classBuilder(genClassName)
+                .addSuperinterface(className(interfaceName))
+                .addModifiers(Modifier.PUBLIC)
+                .addMethod(methodSpec)
+                .build();
         try {
-            FileObject res = filer.createResource(StandardLocation.CLASS_OUTPUT, "", fileName);
-            OutputStream os = res.openOutputStream();
-            BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(os, UTF_8));
-            for (String line : lines) {
-                writer.write(line);
-                writer.newLine();
-            }
-            writer.flush();
-            os.close();
+            JavaFile.builder(Const.GEN_PKG, typeSpec)
+                    .build()
+                    .writeTo(filer);
         } catch (IOException e) {
-            e.printStackTrace();
+            throw new RuntimeException(e);
         }
-    }
 
-    public static String hash(String str) {
-        try {
-            MessageDigest md = MessageDigest.getInstance("MD5");
-            md.update(str.getBytes());
-            return new BigInteger(1, md.digest()).toString(16);
-        } catch (NoSuchAlgorithmException e) {
-            return Integer.toHexString(str.hashCode());
-        }
-    }
+        String fullImplName = Const.GEN_PKG + Const.DOT + genClassName;
+        String className = "ServiceInit" + Const.SPLITTER + hash(genClassName);
 
-    public static boolean notEmpty(String path) {
-        return path != null && path.length() > 0;
-    }
-
-    public static boolean isEmpty(String path) {
-        return path == null || path.length() == 0;
-    }
-
-    public static boolean isEmpty(Collection collection) {
-        return collection == null || collection.isEmpty();
+        new ServiceInitClassBuilder(className)
+                .putDirectly(interfaceName, fullImplName, fullImplName, false)
+                .build();
     }
 
     /**
-     * Returns the binary name of a reference type. For example,
-     * {@code com.google.Foo$Bar}, instead of {@code com.google.Foo.Bar}.
+     * 辅助工具类，用于生成ServiceInitClass，格式如下：
+     * <pre>
+     * package com.sankuai.waimai.router.generated.service;
+     *
+     * import com.sankuai.waimai.router.service.ServiceLoader;
+     *
+     * public class &lt;ClassName&gt; {
+     *     public static void init() {
+     *         ServiceLoader.put(com.xxx.interface1.class, "key1", com.xxx.implementsA.class, false);
+     *         ServiceLoader.put(com.xxx.interface2.class, "key2", com.xxx.implementsB.class, false);
+     *     }
+     * }
+     * </pre>
      */
-    public static String getBinaryName(TypeElement element) {
-        return getBinaryNameImpl(element, element.getSimpleName().toString());
-    }
+    public class ServiceInitClassBuilder {
 
-    private static String getBinaryNameImpl(TypeElement element, String className) {
-        Element enclosingElement = element.getEnclosingElement();
+        private final String className;
+        private final CodeBlock.Builder builder;
+        private final ClassName serviceLoaderClass;
 
-        if (enclosingElement instanceof PackageElement) {
-            PackageElement pkg = (PackageElement) enclosingElement;
-            if (pkg.isUnnamed()) {
-                return className;
-            }
-            return pkg.getQualifiedName() + "." + className;
+        public ServiceInitClassBuilder(String className) {
+            this.className = className;
+            this.builder = CodeBlock.builder();
+            this.serviceLoaderClass = className(Const.SERVICE_LOADER_CLASS);
         }
 
-        TypeElement typeElement = (TypeElement) enclosingElement;
-        return getBinaryNameImpl(typeElement, typeElement.getSimpleName() + "$" + className);
+        public ServiceInitClassBuilder put(String interfaceName, String key, String implementName, boolean singleton) {
+            builder.addStatement("$T.put($T.class, $S, $T.class, $L)",
+                    serviceLoaderClass,
+                    className(interfaceName),
+                    key,
+                    className(implementName),
+                    singleton);
+            return this;
+        }
+
+        public ServiceInitClassBuilder putDirectly(String interfaceName, String key, String implementName, boolean singleton) {
+            // implementName是注解生成的类，直接用$L拼接原始字符串
+            builder.addStatement("$T.put($T.class, $S, $L.class, $L)",
+                    serviceLoaderClass,
+                    className(interfaceName),
+                    key,
+                    implementName,
+                    singleton);
+            return this;
+        }
+
+        public void build() {
+            MethodSpec methodSpec = MethodSpec.methodBuilder(Const.INIT_METHOD)
+                    .addModifiers(Modifier.PUBLIC, Modifier.STATIC)
+                    .returns(TypeName.VOID)
+                    .addCode(this.builder.build())
+                    .build();
+
+            TypeSpec typeSpec = TypeSpec.classBuilder(this.className)
+                    .addModifiers(Modifier.PUBLIC)
+                    .addMethod(methodSpec)
+                    .build();
+            try {
+                JavaFile.builder(Const.GEN_PKG_SERVICE, typeSpec)
+                        .build()
+                        .writeTo(filer);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
     }
 }